refactor: 불필요한 tx proxy 계층 제거
