+++ build.gradle
@@ -10,7 +10,7 @@ repositories {
 
 dependencies {
     implementation 'com.mysql:mysql-connector-j:8.4.0'
-    
+
     testImplementation platform('org.junit:junit-bom:5.11.4')
     testImplementation platform('org.assertj:assertj-bom:3.27.3')
     testImplementation('org.junit.jupiter:junit-jupiter')
