+++ src/main/java/application/JanggiService.java
@@ -1,13 +1,51 @@
 package application;
 
+import model.JanggiGame;
+import model.board.Board;
+import model.board.BoardFactory;
 import model.board.TeamFormation;
+import repository.JanggiRepository;
+import repository.db.TransactionTemplate;
 
-public interface JanggiService {
+public class JanggiService {
 
-    Long createJanggiGame(TeamFormation hanFormation, TeamFormation choFormation);
+    private final TransactionTemplate transactionTemplate;
+    private final JanggiRepository janggiRepository;
 
-    void finishByBigJang(Long janggiId);
+    public JanggiService(TransactionTemplate transactionTemplate, JanggiRepository janggiRepository) {
+        this.transactionTemplate = transactionTemplate;
+        this.janggiRepository = janggiRepository;
+    }
 
-    void movePiece(Long janggiId, MoveCommand command);
+    public Long createJanggiGame(TeamFormation hanFormation, TeamFormation choFormation) {
+        return transactionTemplate.execute(() -> {
+            Board board = BoardFactory.generateDefaultPieces();
+            board.arrangePieces(hanFormation.generate());
+            board.arrangePieces(choFormation.generate());
 
+            JanggiGame janggiGame = new JanggiGame(board);
+            return janggiRepository.save(janggiGame);
+        });
+    }
+
+    public void finishByBigJang(Long janggiId) {
+        transactionTemplate.execute(() -> {
+            JanggiGame janggiGame = getGame(janggiId);
+            janggiGame.finishByBigJang();
+            janggiRepository.update(janggiId, janggiGame);
+        });
+    }
+
+    public void movePiece(Long janggiId, MoveCommand command) {
+        transactionTemplate.execute(() -> {
+            JanggiGame janggiGame = getGame(janggiId);
+            janggiGame.movePiece(command.current(), command.next());
+            janggiRepository.update(janggiId, janggiGame);
+        });
+    }
+
+    private JanggiGame getGame(Long janggiId) {
+        return janggiRepository.findById(janggiId)
+                .orElseThrow(() -> new IllegalArgumentException(janggiId + "번 장기 게임이 존재하지 않습니다."));
+    }
 }
